requirements.txt
--find-links https://k2-fsa.github.io/icefall/piper_phonemize.html

torch
torchaudio
numpy
lhotse
huggingface_hub
safetensors
tensorboard
vocos
pydub
onnxruntime
git+https://github.com/ysharma3501/LinaCodec.git
librosa
transformers<=4.57.6

# Normalization
cn2an
inflect

# Tokenization
jieba
piper_phonemize
pypinyin
setuptools<81

# Configuration
python-dotenv

# API Server
fastapi
uvicorn[standard]
python-multipart
soundfile
